Split vga module
